Images can be now relative paths
In the original slides.md the images had the url
`src=http://localhost:8000/watever`, which is bad because it requires you to
have a local server on port 8000. If you upload the slideshow to some
other server, it won't work.

It would be desirable to have the urls like `src=whatever`, being
`whatever` a relative route to your `index.html`. However, making this
change makes the images to disappear. Why?

The reason is that `ng-bind-html` "sanitizes" the given html before
inserting it, and as part of this sanitization it removes all which does
not fit with its regexps (see [1][]). The solution thus is to force
angular not to sanitize our slides.

This can be done by making the html "trusted" (see [2][])

With these changes it iseven  possible to view the slideshow without a
server. Simply open `index.html` file in the browser.

[1]: http://stackoverflow.com/questions/14363437/ng-bind-html-does-not-load-image-src
[2]: http://stackoverflow.com/questions/19415394/with-ng-bind-html-unsafe-removed-how-do-i-inject-html